Reorganize Code
