--- src/ActionButton.jsx
@@ -1,14 +1,0 @@
-export function ActionButton(props) {
-  return (
-    <button
-      class="action-button"
-      onClick={props.action}
-      disabled={props.disabled}
-    >
-      {props.children}
-    </button>
-  );
-}
-
-/* *** When the button is clicked it should remain disabled *** */
-/* *** until the action is complete. *** */
