+++ src/JS/helperFunctions.js
