--- src/NoteList.css
@@ -1,5 +1,0 @@
-/* button.add-note::before {
-  content: "✏️";
-  margin-right: 0.5rem;
-  cursor: pointer;
-} */
